투표 시 리워드 추가 기능 구현 완료

+++ src/main/java/com/banchan/model/dto/ReviewCountData.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class ReviewCountData {
 
-    private Integer questionId;
+    private Long questionId;
     private Long reviewCount;
 
     public ReviewCountData(Integer questionId, Long reviewCount) {
-        this.questionId = questionId;
+        this.questionId = questionId.longValue();
         this.reviewCount = reviewCount;
     }
 }
